rename previous service objects
