+++ pkg/services/invoice.go
@@ -1,4 +1,4 @@
-package service
+package services
 
 import "net/http"
