+++ pkg/services/invoice/process_invoice.go
@@ -8,15 +8,15 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/transcom/mymove/pkg/db/sequence"
-	"github.com/transcom/mymove/pkg/edi/gex"
 	"github.com/transcom/mymove/pkg/edi/invoice"
 	"github.com/transcom/mymove/pkg/models"
+	"github.com/transcom/mymove/pkg/services"
 )
 
 // ProcessInvoice is a service object to generate/send/record an invoice.
 type ProcessInvoice struct {
 	DB                    *pop.Connection
-	GexSender             gex.SendToGex
+	GexSender             services.SendToGex
 	SendProductionInvoice bool
 	ICNSequencer          sequence.Sequencer
 }
